Antarctic Ice Cores Revised 800KYr CO2 dataset

A dataset containing information of Antarctic ice cores revised 800KYr CO2.

data(ice_core)

Format

A data frame with 1901 rows and 3 variables, each description has 9 components: what, material, error, units, seasonality, archive, detail, method, C or N for Character or Numeric data.

age_gas_calBP

age, gas, , calendar years before present, , , , AICC2012 except for Law Dome WAIS and Siple Dome, N.

co2_ppm

CO2 concentration, , , ppmv, , , corrected, , N

co2_1s_ppm

CO2 concentration sigma mean, , , ppmv, , , Where no individual sigma is given we use average for system/record, , N

Source

ftp://ftp.ncdc.noaa.gov/pub/data/paleo/icecore/antarctica/antarctica2015co2composite.txt

Details

Revised EPICA Dome C and Antarctic composite ice core atmospheric CO2 data. This new version of CO2 composite replaces the old version of Luthi et al. (2008), which contains the analytical bias described in Bereiter et al. 2015 and lower quality data in other sections. For details about the improvements relative to the previous version see supplementary information of Bereiter et al. 2015. For detailed references of all records refer to the supplemetary information of Bereiter et al. 2015. For latest anthropogenic data refer to NOAA/Mauna Loa record. Age unit is in years before present (yr BP) where present refers to 1950 AD. Note, not all records shown in Excel worksheet "all records" are part of the composite.

Notes

Composite built from the following:

-51-1800 yr BP: Law Dome (Rubino et al., 2013)

1.8-2 kyr BP: Law Dome (MacFarling Meure et al., 2006)

2-11 kyr BP: Dome C (Monnin et al., 2001 + 2004)

11-22 kyr BP: WAIS (Marcott et al., 2014) minus 4 ppmv (see text)

22-40 kyr BP: Siple Dome (Ahn et al., 2014)

40-60 kyr BP: TALDICE (Bereiter et al., 2012)

60-115 kyr BP: EDML (Bereiter et al., 2012)

105-155 kyr BP: Dome C Sublimation (Schneider et al., 2013)

155-393 kyr BP: Vostok (Petit et al., 1999)

393-611 kyr BP: Dome C (Siegenthaler et al., 2005)

612-800 kyr BP: Dome C (Bereiter et al., 2014)

If millennial scale or smaller details of the composite are studied, we recommend to look into all records available for that period and not only in the composite.

References

Bereiter, B., Eggleston, S., Schmitt, J., Nehrbass‐Ahles, C., Stocker, T. F., Fischer, H., Kipfstuhl, S., and Chappellaz, J. (2015), Revision of the EPICA Dome C CO2 record from 800 to 600 kyr before present, Geophys. Res. Lett., 42, 542– 549, doi:10.1002/2014GL061957.

Author

Bernhard Bereiter bbereiter@ucsd.edu

Sarah Eggleston

Jochen Schmitt

Christoph Nehrbass-Ahles

Thomas F. Stocker

Hubertus Fischer

Sepp Kipfstuhl

Jerome Chappellaz