CRU TS 4.04 Landmask dataset

A dataset containing information of land mask with a grid of 0.5 by 0.5 degrees resolution. TRUE values represent a land grid cell.

data(land_mask)

Format

A 2D matrix of 720 x 360 (lon x lat) with boolean values:

TRUE

Land grid cell.

FALSE

Non-land grid cell.

Details

The mask was generated by looking at the values of tmn and vap, for each grid cell, if at least one timestep had a non-missing value, then the entire grid cell was marked as a land grid cell.

References

University of East Anglia Climatic Research Unit; Harris, I.C.; Jones, P.D.; Osborn, T. (2020): CRU TS4.04: Climatic Research Unit (CRU) Time-Series (TS) version 4.04 of high-resolution gridded data of month-by-month variation in climate (Jan. 1901- Dec. 2019). Centre for Environmental Data Analysis. https://catalogue.ceda.ac.uk/uuid/89e1e34ec3554dc98594a5732622bce9