P-model inverter class

This class allows to create objects of the P-model.

Methods

Public methods

P_model_inverter$new()

P_model_inverter$calculate_m_true()

Method new()

Create a new p-model object.

Usage

P_model_inverter$new(T_diff, T_ref, m_rec, c_ratio, lat = -30, ...)

Arguments

T_diff

Temperature difference.

T_ref

Reference temperature (e.g. modern temperature).

m_rec

Reconstructed moisture index.

c_ratio

CO2 ratio.

lat

Latitude.

...

Optional parameters (not used).

Method calculate_m_true()

Calculate true moisture index.

Usage

P_model_inverter$calculate_m_true()

Returns

A list with three elements:

mi:

Numeric value of moisture index.

cph:

Boolean flat to indicate whether or not the compensation point 'law' is upheld.

ci:

Numeric value of c_i.

Examples

codos::P_model_inverter$new(T_diff = 1.334567,
                            T_ref = 11.57957,
                            m_rec = 0.3357231,
                            c_ratio = 0.7361765,
                            lat = -30)$calculate_m_true()

#> $mi
#> [1] 0.6402775
#> 
#> $cph
#> [1] TRUE
#> 
#> $ci
#> [1] 41.56653
#>